requirements.txt
alembic==1.4.0
aniso8601==6.0.0
appdirs==1.4.3
astroid==2.2.5
atomicwrites==1.3.0
attrs==19.3.0
auth0-python==3.9.1
Authlib==0.14.1
bcrypt==3.1.7
blinker==1.4
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
Click==7.0
colorama==0.4.3
cryptography==2.8
distlib==0.3.0
ecdsa==0.13.2
filelock==3.0.12
Flask==1.0.2
Flask-Bcrypt==0.7.1
Flask-Cors==3.0.7
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==2.5.2
Flask-RESTful==0.3.7
Flask-SQLAlchemy==2.4.0
Flask-WTF==0.14.3
future==0.17.1
gunicorn==20.0.4
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2020.3.12
httpx==0.12.0
hyperframe==5.2.0
idna==2.9
importlib-metadata==1.5.0
isort==4.3.18
itsdangerous==1.1.0
Jinja2==2.10.1
lazy-object-proxy==1.4.0
Mako==1.1.1
MarkupSafe==1.1.1
mccabe==0.6.1
mock==4.0.1
more-itertools==8.2.0
packaging==20.3
passlib==1.7.2
pipenv==2018.11.26
pluggy==0.13.1
psycopg2-binary==2.8.2
py==1.8.1
pyasn1==0.4.8
pycparser==2.19
pyparsing==2.4.6
pytest==5.3.5
python-dateutil==2.8.1
python-dotenv==0.12.0
python-editor==1.0.4
python-jose==3.1.0
pytz==2019.1
requests==2.23.0
rfc3986==1.3.2
rsa==4.0
six==1.12.0
sniffio==1.1.0
SQLAlchemy==1.3.4
typed-ast==1.3.5
urllib3==1.25.8
virtualenv==20.0.13
virtualenv-clone==0.5.3
wcwidth==0.1.8
Werkzeug==0.15.2
wrapt==1.11.1
WTForms==2.2.1
zipp==3.1.0
